fix(server): widen error columns to BIGINT to handle XB10 overflow
XB10 modems report total_corrected_errors values (~2.99B) that exceed
PostgreSQL's 32-bit INTEGER max (2,147,483,647). asyncpg was rejecting
these writes and surfacing a misleading 400 "Database error" to clients.

- Widen uptime_seconds, total_corrected_errors, total_uncorrected_errors
  to BigInteger in modem_check.py
- Add structured error logging in upload_service.py for future DB errors
- Add regression tests in test_upload.py and test_metric_extraction.py

DB migration (INTEGER → BIGINT) already applied to prod and test DBs.

+++ cloudserver/app/models/modem_check.py
@@ -1,7 +1,7 @@
 """
 Modem Check model for storing cable modem diagnostic data.
 """
-from sqlalchemy import Column, Integer, String, Float, DateTime, Text, Index, desc
+from sqlalchemy import Column, Integer, BigInteger, String, Float, DateTime, Text, Index, desc
 from sqlalchemy.dialects.postgresql import JSONB
 
 from app.core.database import Base
@@ -25,15 +25,15 @@ class ModemCheck(Base):
 
     # System info
     firmware = Column(String(255), nullable=True)
-    uptime_seconds = Column(Integer, nullable=True)
+    uptime_seconds = Column(BigInteger, nullable=True)
     system_time = Column(DateTime, nullable=True)
 
     # Signal quality metrics (for quick queries)
     avg_downstream_power = Column(Float, nullable=True)
     avg_downstream_snr = Column(Float, nullable=True)
     avg_upstream_power = Column(Float, nullable=True)
-    total_corrected_errors = Column(Integer, nullable=True)
-    total_uncorrected_errors = Column(Integer, nullable=True)
+    total_corrected_errors = Column(BigInteger, nullable=True)
+    total_uncorrected_errors = Column(BigInteger, nullable=True)
 
     # Speed test results
     speedtest_enabled = Column(Integer, nullable=True)  # 0=disabled, 1=enabled
